strip trailing newlines in python version string
Did this by replacing the deprecated exec_program with its successor execute_process, which includes an option for stripping whitespace in output variables.

Note that the newlines I got may have been due to an old ROOT version or due to some other platform (macOS) or environment dependent settings, but still, this makes the code a bit more robust to such variations.